+++ src/commands/plugins.ts
@@ -16,11 +16,7 @@ export async function plugins() {
         await download(url);
     }
     const pluginNames = e.map((e) => e.split('/').pop());
-    console.log(
-        `Successfully downloaded plugin(s):\n${greenBright(
-            pluginNames.join('\n')
-        )}`
-    );
+    console.log(`Successfully downloaded plugin(s):\n${greenBright(pluginNames.join('\n'))}`);
 }
 
 async function download(url: string) {