+++ src/index.ts
@@ -34,23 +34,22 @@ program
     .description('Easy way to add extra things in your sern project')
     .action((...args) => importDynamic('extra').then(m => m.extra(...args)));
 
-program
-    .command('publish')
-    .description('Manage your slash commands')
-    .option('-a, --all', 'Publish all commands')
-    .option('-t, --token [token]')
-    .option('--appId [applicationId]')
-    .argument(
-        '[pattern]',
-        'glob pattern that will locate all published files',
-        '<<none>>'
-    )
-    .action(async (...args) => importDynamic('publish.js').then(m => m.publish(...args)));
+program //
+    .command('commands')
+    .description('Defacto way to manage your slash commands')
+    .addCommand(
+        new Command('publish')
+            .description('New way to manage your slash commands')
+            .option('-i, --import [scriptPath...]', 'Prerequire a script to load into publisher')
+            .option('-t, --token [token]')
+            .option('--appId [applicationId]')
+            .argument('[path]', 'path with respect to current working directory that will locate all published files')
+            .action(async (...args) => importDynamic('publish.js').then(m => m.publish(...args)))
+    );
 
 program 
     .command('build')
     .description('Build your bot')
     .action(async (...args) => importDynamic('build.js').then(m => m.build(...args)))
-
 
 program.parse();