+++ tsup.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsup';
 import { esbuildPluginVersionInjector } from 'esbuild-plugin-version-injector';
 const shared = {
-    entry: ['src/index.ts', 'src/create-publish.mts', 'src/loader.ts', 'src/commands/**'],
+    entry: ['src/index.ts', 'src/create-publish.mts', 'src/commands/**'],
     clean: true,
     sourcemap: true,
 };